g-s: updated structure
